+++ src/components/Profile/FollowButton/FollowButton.tsx
@@ -2,6 +2,8 @@ import { MutatorOptions } from 'swr';
 import { FOLLOW_API } from 'utils/api';
 import axios from 'axios';
 import * as S from './styles';
+import { useFollow } from 'hooks/useFollow';
+import { useEffect } from 'react';
 
 interface Props {
   mutate: (
@@ -13,6 +15,8 @@ interface Props {
 }
 
 export const FollowButton = ({ mutate, id, page }: Props) => {
+  const { isFollowed, setIsFollowed, checkFollow, loading } = useFollow();
+
   const addFollowing = (id: number) => {
     const config = {
       method: 'post',
@@ -27,6 +31,7 @@ export const FollowButton = ({ mutate, id, page }: Props) => {
         if (!(response.data.message === '팔로우 생성 성공')) {
           return;
         }
+        setIsFollowed(true);
         mutate();
       })
       .catch(error => alert(error.response.data.message));
@@ -45,9 +50,18 @@ export const FollowButton = ({ mutate, id, page }: Props) => {
       .catch(error => console.log(error));
   };
 
+  useEffect(() => {
+    checkFollow(id);
+  }, [checkFollow, id]);
+
   return page === 'following' ? (
     <S.Button onClick={() => deleteFollowing(id)}>언팔로우</S.Button>
-  ) : (
-    <S.Button onClick={() => addFollowing(id)}>팔로우</S.Button>
-  );
+  ) : !isFollowed ? (
+    <S.Button
+      onClick={() => addFollowing(id)}
+      style={{ display: loading ? 'none' : 'block' }}
+    >
+      팔로우
+    </S.Button>
+  ) : null;
 };
